Moved the Config class to BottomBarTab, since it's more logical.
